# docs/phase-1-api-contracts/openapi.yaml
...
openapi: 3.1.0
info:
  title: RAG Refinement System API
  version: 1.0.0
  summary: TOC-routed retrieval-refinement layer — routing-only enterprise API plus a streaming personal-tool Q&A surface.
  description: |
    # RAG Refinement System API

    A document-structure-aware retrieval layer that selects relevant document
    sections (via a TOC-based Router Agent) **before** vector retrieval. Two
    product surfaces share one contract (HLD §1.1):

    - **Enterprise Refinement API** — `POST /v1/route` is **routing-only**:
      it returns the optimal retrieval scope `{relevant_sections[], page_ranges[],
      confidence[], fallback, ...}` and **never calls the generation LLM**
      (binding AGREED CONTRACT: ai-engineer <-> python-backend-engineer).
      Authenticated with an API key (ADR-7).
    - **Personal Tool** — `POST /v1/answer` streams a generated, cited answer
      token-by-token over Server-Sent Events (SSE), ending with a final event
      carrying `{answer, citations[], routing{}}`. Authenticated with an
      OAuth2 / JWT bearer token (ADR-7).

    `section_id` is the universal join/filter key across the Document -> Section
    -> Chunk hierarchy (HLD §4, §7.5).

    ## Versioning strategy
    The API is versioned in the **URL path** (`/v1/...`). The version increments
    on breaking changes only (per global API standard: version APIs in the URL
    path). Backwards-compatible additions (new optional fields, new endpoints,
...

    ## Deprecation policy
...

...

...
servers:
...
paths:
...
  /v1/answer:
    post:
      tags: [Answers]
      operationId: answerQuery
      summary: Stream a generated, cited answer over Server-Sent Events (personal tool).
      description: |
        Personal-tool endpoint. Routes the query, performs targeted retrieval,
        optionally reranks, then **streams** the generated answer token-by-token
        as Server-Sent Events (`text/event-stream`). The stream emits:

        - repeated `event: token` messages whose `data` is a `TokenEvent`
          (incremental answer text);
        - a terminal `event: final` message whose `data` is an `AnswerFinalEvent`
          carrying `{answer, citations[], routing{}}`;
        - on error, an `event: error` message whose `data` is a `Problem`.

        A non-streaming error (auth, validation, not-found) is returned as a
        normal RFC-7807 response **before** the stream begins. Errors that occur
        mid-stream are delivered as an SSE `error` event.
      x-fr: [FR-007, FR-008, FR-009, FR-011, FR-012, FR-018, FR-016]
      x-audit-log: true
      x-rate-limit:
        scope: jwt-subject
        free: 50/month
        pro: 1000/month
        power: 10000/month
        burst: 5/second
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnswerRequest'
      responses:
        '200':
          description: |
            SSE stream opened. The body is `text/event-stream`. See the schemas
            `TokenEvent`, `AnswerFinalEvent`, and `Problem` for the `data`
            payloads of the `token`, `final`, and `error` events respectively.
          headers:
            Cache-Control:
              schema: { type: string, const: 'no-cache' }
            Connection:
              schema: { type: string, const: 'keep-alive' }
          content:
            text/event-stream:
              schema:
                $ref: '#/components/schemas/SseAnswerStream'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/DocumentNotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: Enterprise API key (with rotation, ADR-7 / FR-010 / NFR-015).
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OAuth2 / JWT bearer token for the personal tool (ADR-7 / FR-007 / NFR-015).
    oauth2:
      type: oauth2
      description: OAuth2 authorization-code flow backing the personal-tool JWT (ADR-7).
      flows:
        authorizationCode:
          authorizationUrl: https://auth.rag-refinement.example.com/oauth2/authorize
          tokenUrl: https://auth.rag-refinement.example.com/oauth2/token
          refreshUrl: https://auth.rag-refinement.example.com/oauth2/token
          scopes:
            documents:read: Read documents and TOCs.
            documents:write: Upload and delete documents.
            answers:read: Stream answers.
...
  responses:
    BadRequest:
      description: Malformed request.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: https://api.rag-refinement.example.com/problems/bad-request
            title: Bad Request
            status: 400
            code: BAD_REQUEST
            detail: The request body could not be parsed as JSON.
    Unauthorized:
      description: Missing or invalid credentials.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: https://api.rag-refinement.example.com/problems/unauthorized
            title: Unauthorized
            status: 401
            code: UNAUTHORIZED
            detail: API key or bearer token is missing or invalid.
...
    DocumentNotFound:
      description: Document does not exist (or has been erased).
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: https://api.rag-refinement.example.com/problems/document-not-found
            title: Not Found
            status: 404
            code: DOCUMENT_NOT_FOUND
            detail: No document with the given id exists.
    ValidationError:
      description: Request body failed schema validation.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ValidationProblem' }
          example:
            type: https://api.rag-refinement.example.com/problems/validation-error
            title: Unprocessable Entity
            status: 422
            code: VALIDATION_ERROR
            detail: One or more fields failed validation.
            errors:
              - field: confidence_threshold
                message: must be between 0.0 and 1.0
...
    RateLimited:
      description: Rate limit exceeded.
      headers:
        Retry-After:
          description: Seconds until the caller may retry.
          schema: { type: integer, minimum: 0 }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: https://api.rag-refinement.example.com/problems/rate-limited
            title: Too Many Requests
            status: 429
            code: RATE_LIMITED
            detail: The rate limit for this API key has been exceeded.
    InternalError:
      description: Unexpected server error.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: https://api.rag-refinement.example.com/problems/internal-error
            title: Internal Server Error
            status: 500
            code: INTERNAL_ERROR
            detail: An unexpected error occurred. The incident has been logged.
    ServiceUnavailable:
      description: A required dependency is unavailable; routing may degrade to fallback.
      headers:
        Retry-After:
          schema: { type: integer, minimum: 0 }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: https://api.rag-refinement.example.com/problems/service-unavailable
            title: Service Unavailable
            status: 503
            code: SERVICE_UNAVAILABLE
            detail: A required dependency (vector DB or LLM provider) is unreachable.
  schemas:
    DocumentId:
      type: string
      pattern: '^doc_[A-Za-z0-9]{6,}$'
      examples: [doc_abc123]
      description: Stable document identifier.
    SectionId:
      type: string
      pattern: '^sec_[A-Za-z0-9]{1,}$'
      examples: [sec_warranty]
      description: Universal section identifier — the join/filter key (HLD §7.5).
    QueryId:
      type: string
      pattern: '^qry_[A-Za-z0-9]{1,}$'
      examples: [qry_xyz789]
      description: Correlation id for a routing/answer request (NFR-009 tracing).
    Confidence:
      type: number
      format: float
      minimum: 0.0
      maximum: 1.0
      description: Router confidence that a section is relevant (PRD 8.3).
...
    AnswerRequest:
      type: object
      additionalProperties: false
      required: [document_id, query]
      properties:
        document_id: { $ref: '#/components/schemas/DocumentId' }
        query:
          type: string
          minLength: 1
          maxLength: 4000
          description: The user's natural-language question (untrusted input).
          x-pii: true
        confidence_threshold:
          allOf: [ { $ref: '#/components/schemas/Confidence' } ]
          default: 0.7
        max_sections:
          type: integer
          minimum: 1
          maximum: 20
          default: 3
        rerank:
          type: boolean
          default: false
          description: Apply Cohere Rerank-3 post-route (FR-016, ADR-8).
        no_retention:
          type: boolean
          default: false
          description: Purge any per-request artifacts after the answer (FR-027).
    Citation:
      type: object
      additionalProperties: false
      required: [section_title, page_start, page_end]
      properties:
        section_id: { $ref: '#/components/schemas/SectionId' }
        section_title:
          type: string
          x-pii: true
        page_start: { type: integer, minimum: 1 }
        page_end: { type: integer, minimum: 1 }
    RoutingSummary:
      type: object
      additionalProperties: false
      required: [sections, confidence, fallback]
      properties:
        sections:
          type: array
          items: { $ref: '#/components/schemas/SectionId' }
        confidence:
          type: array
          items: { $ref: '#/components/schemas/Confidence' }
        fallback: { type: boolean }
        rationale:
          type: string
          description: Interpretability rationale (FR-012, NFR-011).
    TokenEvent:
      type: object
      additionalProperties: false
      required: [token]
      description: "SSE `event: token` data payload - an incremental answer fragment."
      properties:
        query_id: { $ref: '#/components/schemas/QueryId' }
        token:
          type: string
          x-pii: true
          description: A fragment of the streamed answer text.
    AnswerFinalEvent:
      type: object
      additionalProperties: false
      required: [answer, citations, routing]
      description: "SSE `event: final` data payload - the completed cited answer."
      properties:
        query_id: { $ref: '#/components/schemas/QueryId' }
        answer:
          type: string
          x-pii: true
          description: The complete generated answer.
        citations:
          type: array
          items: { $ref: '#/components/schemas/Citation' }
        routing:
          $ref: '#/components/schemas/RoutingSummary'
    SseAnswerStream:
      type: object
      description: |
        Documentation-only envelope describing the union of SSE `data` payloads
        on the `/v1/answer` stream. The wire format is `text/event-stream`; each
        message's `data` is one of these JSON objects keyed by its `event` name
        (`token` -> TokenEvent, `final` -> AnswerFinalEvent, `error` -> Problem).
      oneOf:
        - $ref: '#/components/schemas/TokenEvent'
        - $ref: '#/components/schemas/AnswerFinalEvent'
        - $ref: '#/components/schemas/Problem'
...
    Problem:
      type: object
      description: RFC 7807 problem detail (application/problem+json).
      additionalProperties: true
      required: [type, title, status, code]
      properties:
        type:
          type: string
          format: uri
          description: A URI identifying the problem type.
        title:
          type: string
          description: Short, human-readable summary of the problem type.
        status:
          type: integer
          minimum: 100
          maximum: 599
          description: HTTP status code.
        code:
          type: string
          description: Machine-readable error code (see error_catalog.json).
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence.
        instance:
          type: string
          format: uri
          description: URI reference identifying this occurrence.
        query_id:
          $ref: '#/components/schemas/QueryId'
    ValidationProblem:
      allOf:
        - $ref: '#/components/schemas/Problem'
        - type: object
          properties:
            errors:
              type: array
              items:
                type: object
                additionalProperties: false
                required: [field, message]
                properties:
                  field: { type: string }
                  message: { type: string }
security:
  - apiKey: []
  - bearerAuth: []